requirements.txt
Django==3.2.12
python-rapidjson==1.6
djangorestframework==3.13.1